requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pytest
pytest
httpx
python-dotenv
